Create earth bay
Welcome to Earth Bay! Our mission is to create a unique Bio Art Village in Vengurla that showcases the best of what the region has to offer. We are committed to providing upscale services at a fair price, improving local infrastructure and hotel services, and offering a wide range of recreational activities for all ages. We follow India's Tourism Policy and strive to extend the tourist season beyond the low monsoon period. Thank you for visiting our website and we hope to see you at Earth Bay soon!